Fixes performance issue with leaderboards#show
Performance issue existed where leaderboards show would take an
incredible amount of time if the user had not had any achievements due
to a full scan and sort of the achievements table. Prevented with an
achievements.empty? call before running the query that could potentially
take forever.

Also reduces the amount of data sent whenever an entry and medal
statistic are sent to the client.