Adds MemoryCache for serializers
Now has a small MemoryStore cache to store the serializer objects in so
that we don't have to regenerate all of the objects on every change.
This adds a little bit better performance, but nothing crazy.

Fixes bug on realtime leaderboard where it did not unsubscribe from not
shown leaderboards. Deletes some dead code there too.